+++ examples/cli.py
@@ -1,6 +1,6 @@
 import click
 
-from examples.ds.plots import plots_binning, plots_running
+from examples.ds.plots import plots_binning, plots_running, plots_smoothing
 
 
 @click.group()
@@ -9,6 +9,7 @@ def cli():
 
 cli.add_command(plots_binning)
 cli.add_command(plots_running)
+cli.add_command(plots_smoothing)
 
 
 if __name__ == "__main__":