+++ examples/ds/plots.py
@@ -6,7 +6,7 @@ import numpy as np
 from pelutils.ds.plots import (
     linear_binning, log_binning, normal_binning, get_bins,
     figsize_wide, rc_params, rc_params_small, update_rc_params,
-    running_avg,
+    running_avg, exp_running_avg, running_avg_smoothing, tab_colours
 )
 from pelutils.ds.distributions import norm, lognorm
 
@@ -65,17 +65,75 @@ def plots_running():
 
     # Plot data with running average function and few neighbors
     plt.subplot(221)
-    plt.plot(x, y)
-    plt.plot(*running_avg(x, y, neighbors=1))
-    plt.title("Running avg., low smoothing")
+    plt.scatter(x, y)
+    plt.plot(*running_avg(x, y, neighbors=1), c=tab_colours[1])
+    plt.title("Running avg., n=1")
     plt.grid()
 
     # Same but with higher smoothing
     plt.subplot(222)
-    plt.plot(x, y)
-    plt.plot(*running_avg(x, y, neighbors=4))
-    plt.title("Running avg., high smoothing")
+    plt.scatter(x, y)
+    plt.plot(*running_avg(x, y, neighbors=4), c=tab_colours[1])
+    plt.title("Running avg., n=4")
     plt.grid()
 
-    # plt.tight_layout()
+    # Same but with higher smoothing
+    plt.subplot(223)
+    plt.scatter(x, y)
+    plt.plot(*exp_running_avg(x, y), c=tab_colours[1])
+    plt.title(r"Exp. running avg., $\alpha=0.2$")
+    plt.grid()
+
+    # Same but with higher smoothing
+    plt.subplot(224)
+    plt.scatter(x, y)
+    plt.plot(*exp_running_avg(x, y, reverse=True), c=tab_colours[1])
+    plt.title(r"Reverse exp. running avg., $\alpha=0.2$")
+    plt.grid()
+
+    plt.show()
+
+
+@click.command("plots-smoothing")
+def plots_smoothing():
+    update_rc_params(rc_params_small)
+    plt.figure(figsize=(30, 20))
+
+    # Generate noisy data
+    n = 100
+    x = np.linspace(-3, 4, n)
+    y = np.sin(x)
+    y += np.random.randn(n) / 3
+    subsample = np.random.randint(0, 2, n).astype(bool)
+    subsample[n//2-10:n//2+5] = False
+    x, y = x[subsample], y[subsample]
+
+    # Plot data with running average function and few neighbors
+    plt.subplot(221)
+    plt.scatter(x, y)
+    plt.plot(*running_avg(x, y, neighbors=2), c=tab_colours[1])
+    plt.title("Running avg., n=2")
+    plt.grid()
+
+    # Plot data with running average function and few neighbors
+    plt.subplot(222)
+    plt.scatter(x, y)
+    plt.plot(*running_avg(x, y, neighbors=4), c=tab_colours[1])
+    plt.title("Running avg., n=4")
+    plt.grid()
+
+    # Running avg. with smoothing
+    plt.subplot(223)
+    plt.scatter(x, y)
+    plt.plot(*running_avg_smoothing(x, y, neighbors=12, samples=300), c=tab_colours[1])
+    plt.title("With smoothing, n=12, samples=300")
+    plt.grid()
+
+    # Running avg. with more smoothing
+    plt.subplot(224)
+    plt.scatter(x, y)
+    plt.plot(*running_avg_smoothing(x, y, neighbors=20, samples=300), c=tab_colours[1])
+    plt.title("With smoothing, n=20, samples=300")
+    plt.grid()
+
     plt.show()